update: add new confirmation delete

+++ composables/useList.ts
@@ -4,10 +4,13 @@
 export const useList = () => {
   async function destroy(id: string, onDestroy?: () => void) {
     try {
+      const res = confirm("Are you sure you want to destroy?")
+      if (res) {  // jika user menekan tombol ok, maka proses delete dijalanka
         await useFetch(`/api/lists/${id}`, {
           method: "DELETE",
         });
         onDestroy?.();
+      }
     } catch (e: any) {
       ElNotification({
         title: "Error",
